backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and caching
redis==5.0.1
hiredis==2.2.3; platform_system != "Windows"  # Optional, requires C++ build tools on Windows

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-dotenv==1.0.0

# Email integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
msal==1.25.0
imapclient==2.3.1

# Document processing and OCR
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2

# AI/ML
spacy==3.7.2
sentence-transformers==2.2.2
transformers==4.35.2
torch>=2.2.0; platform_system != "Windows"  # Optional, only for advanced ML features
ollama==0.1.6
httpx==0.25.2

# Async task processing
celery==5.3.4
flower==2.0.1

# AWS
boto3==1.29.7
botocore==1.32.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator>=2.1.1
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
